+++ libs/cardano-ledger-core/src/Cardano/Ledger/Core.hs
@@ -43,6 +43,10 @@ module Cardano.Ledger.Core (
   EraPParams (..),
   mkCoinTxOut,
   wireSizeTxF,
+  binaryUpgradeTx,
+  binaryUpgradeTxBody,
+  binaryUpgradeTxWits,
+  binaryUpgradeTxAuxData,
 
   -- * Era
   module Cardano.Ledger.Core.Era,
@@ -76,6 +80,7 @@ import Cardano.Ledger.Binary (
   Annotator,
   DecCBOR,
   DecShareCBOR (Share),
+  DecoderError,
   EncCBOR (..),
   EncCBORGroup,
   Interns,
@@ -85,6 +90,7 @@ import Cardano.Ledger.Binary (
   mkSized,
   serialize,
   serialize',
+  translateViaCBORAnnotator,
  )
 import Cardano.Ledger.Coin (Coin)
 import Cardano.Ledger.Compactible (Compactible (..))
@@ -102,6 +108,7 @@ import Cardano.Ledger.Rewards (Reward (..), RewardType (..))
 import Cardano.Ledger.TxIn (TxId (..), TxIn (..))
 import Cardano.Ledger.Val (Val (..), inject)
 import Control.DeepSeq (NFData)
+import Control.Monad.Except (Except)
 import Data.Aeson (ToJSON)
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Lazy as LBS
@@ -114,7 +121,7 @@ import Data.MemPack
 import Data.Sequence.Strict (StrictSeq)
 import Data.Set (Set)
 import qualified Data.Set as Set
-import Data.Void (Void)
+import Data.Text (Text)
 import Data.Word (Word32, Word64)
 import GHC.Stack (HasCallStack)
 import Lens.Micro
@@ -138,9 +145,6 @@ class
   EraTx era
   where
   type Tx era = (r :: Type) | r -> era
-
-  type TxUpgradeError era :: Type
-  type TxUpgradeError era = Void
 
   mkBasicTx :: TxBody era -> Tx era
 
@@ -174,11 +178,6 @@ class
     Int ->
     Coin
 
-  upgradeTx ::
-    EraTx (PreviousEra era) =>
-    Tx (PreviousEra era) ->
-    Either (TxUpgradeError era) (Tx era)
-
 class
   ( EraTxOut era
   , EraTxCert era
@@ -197,9 +196,6 @@ class
   where
   -- | The body of a transaction.
   data TxBody era
-
-  type TxBodyUpgradeError era :: Type
-  type TxBodyUpgradeError era = Void
 
   mkBasicTxBody :: TxBody era
 
@@ -255,22 +251,6 @@ class
   -- tooling to figure out how many witnesses should be supplied for Genesis keys.
   getGenesisKeyHashCountTxBody :: TxBody era -> Int
   getGenesisKeyHashCountTxBody _ = 0
-
-  -- | Upgrade the transaction body from the previous era.
-  --
-  -- This can fail where elements of the transaction body are deprecated.
-  -- Compare this to `translateEraThroughCBOR`:
-  -- - `upgradeTxBody` will use the Haskell representation, but will not
-  --   preserve the serialised form. However, it will be suitable for iterated
-  --   translation through eras.
-  -- - `translateEraThroughCBOR` will preserve the binary representation, but is
-  --   not guaranteed to work through multiple eras - that is, the serialised
-  --   representation from era n is guaranteed valid in era n + 1, but not
-  --   necessarily in era n + 2.
-  upgradeTxBody ::
-    EraTxBody (PreviousEra era) =>
-    TxBody (PreviousEra era) ->
-    Either (TxBodyUpgradeError era) (TxBody era)
 
 -- | Abstract interface into specific fields of a `TxOut`
 class
@@ -463,13 +443,6 @@ class
 
   metadataTxAuxDataL :: Lens' (TxAuxData era) (Map Word64 Metadatum)
 
-  -- | Every era, except Shelley, must be able to upgrade a `TxAuxData` from a previous
-  -- era.
-  --
-  -- /Warning/ - Important to note that any memoized binary representation will not be
-  -- preserved. If you need to retain underlying bytes you can use `translateEraThroughCBOR`
-  upgradeTxAuxData :: EraTxAuxData (PreviousEra era) => TxAuxData (PreviousEra era) -> TxAuxData era
-
   validateTxAuxData :: ProtVer -> TxAuxData era -> Bool
 
 -- | Compute a hash of `TxAuxData`
@@ -500,8 +473,6 @@ class
   bootAddrTxWitsL :: Lens' (TxWits era) (Set BootstrapWitness)
 
   scriptTxWitsL :: Lens' (TxWits era) (Map ScriptHash (Script era))
-
-  upgradeTxWits :: EraTxWits (PreviousEra era) => TxWits (PreviousEra era) -> TxWits era
 
 -- | This is a helper lens that will hash the scripts when adding as witnesses.
 hashScriptTxWitsL ::
@@ -648,3 +619,43 @@ wireSizeTxF =
         then fromIntegral n
         else error $ "Impossible: Size of the transaction is too big: " ++ show n
 {-# INLINEABLE wireSizeTxF #-}
+
+-- | Translate a transaction through its binary representation from previous to current era.
+binaryUpgradeTx ::
+  forall era.
+  (Era era, ToCBOR (Tx (PreviousEra era)), DecCBOR (Annotator (Tx era))) =>
+  -- | Label for error reporting
+  Text ->
+  Tx (PreviousEra era) ->
+  Except DecoderError (Tx era)
+binaryUpgradeTx msg tx = translateViaCBORAnnotator (eraProtVerLow @era) msg tx
+
+-- | Translate a tx body through its binary representation from previous to current era.
+binaryUpgradeTxBody ::
+  forall era.
+  (Era era, ToCBOR (TxBody (PreviousEra era)), DecCBOR (Annotator (TxBody era))) =>
+  -- | Label for error reporting
+  Text ->
+  TxBody (PreviousEra era) ->
+  Except DecoderError (TxBody era)
+binaryUpgradeTxBody msg txBody = translateViaCBORAnnotator (eraProtVerLow @era) msg txBody
+
+-- | Translate tx witnesses through its binary representation from previous to current era.
+binaryUpgradeTxWits ::
+  forall era.
+  (Era era, ToCBOR (TxWits (PreviousEra era)), DecCBOR (Annotator (TxWits era))) =>
+  -- | Label for error reporting
+  Text ->
+  TxWits (PreviousEra era) ->
+  Except DecoderError (TxWits era)
+binaryUpgradeTxWits msg txWits = translateViaCBORAnnotator (eraProtVerLow @era) msg txWits
+
+-- | Translate tx auxData through its binary representation from previous to current era.
+binaryUpgradeTxAuxData ::
+  forall era.
+  (Era era, ToCBOR (TxAuxData (PreviousEra era)), DecCBOR (Annotator (TxAuxData era))) =>
+  -- | Label for error reporting
+  Text ->
+  TxAuxData (PreviousEra era) ->
+  Except DecoderError (TxAuxData era)
+binaryUpgradeTxAuxData msg txAuxData = translateViaCBORAnnotator (eraProtVerLow @era) msg txAuxData